+++ src/senaite/astm/adapters/spotchem/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import se1520  # noqa: F401